+++ BeastsOfBurden/CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2021-03-19
+### Changed 
+* Cart detach range and attachments offset are now based on the radius of the animal it is being attached to, this was to help make the mod more compatible with mods like All Tameable. Detach distance for Wolf, Boar, and Lox was replaced with Detach Distance Factor which is used for all creatures.
+* attachToOtherTamed setting now defaults to True. 
+* Removed logging when character type was other than Player, Boar, Wolf, or Lox. 
+
+### Removed
+* Configuration Settings for 
+
 ## [1.0.2] - 2021-03-18
 ### Fixed
 * Cart detaches when dodging or teleporting as in vanilla Valheim
@@ -13,14 +22,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Ability to configure which animal types can connect to cart
 
-
 ### Changed
 * Set default cart offset for unknown animal types to be the same as used for Wolves and Boars
 
 ## [1.0.1] - 2021-03-17
 ### Fixed
 * NullReferenceException when attaching cart to player. I wasn't able to reproduce this but I added some more aggressive null checks and cart now attaches based on Character transform not Tameable transform. Hopefully this will make it more friendly with other mods. Added additional logging around this as well. 
-
 
 ### Added
 * This changelog :-)